Open telemetry observability phase 1
This commit was very hard to get working. I ended up also chatting with
the Grafana Cloud support bot to do a bunch of debugging. It seems that
there may never have been a problem other than the sample data that
github copilot having timestamps that are old enough that grafana was
dropping them when it received them. The Grafana AI also initially
supplied a curl command with the same problem.

The two scripts came from the Grafana AI, as did the 12345 port number
in grafana-alloy/config.alloy